+++ R/RcppExports.R
@@ -37,15 +37,15 @@ d2elldbeta2_pl_i_fun_exact <- function(beta_pl, c_pl, x_pl_ext_i, y_i) {
     .Call(`_evinf_d2elldbeta2_pl_i_fun_exact`, beta_pl, c_pl, x_pl_ext_i, y_i)
 }
 
-log_lik_fun <- function(gamma_z, gamma_pl, beta_nb, alpha_nb, beta_pl, c_pl, x_mult_z_ext, x_mult_pl_ext, x_nb_ext, x_pl_ext, y, offset_nb, offset_zc, offset_pl_mult) {
-    .Call(`_evinf_log_lik_fun`, gamma_z, gamma_pl, beta_nb, alpha_nb, beta_pl, c_pl, x_mult_z_ext, x_mult_pl_ext, x_nb_ext, x_pl_ext, y, offset_nb, offset_zc, offset_pl_mult)
+log_lik_fun <- function(gamma_z, gamma_pl, beta_nb, alpha_nb, beta_pl, c_pl, x_mult_z_ext, x_mult_pl_ext, x_nb_ext, x_pl_ext, y, offset_nb, offset_zc, offset_pl_mult, w) {
+    .Call(`_evinf_log_lik_fun`, gamma_z, gamma_pl, beta_nb, alpha_nb, beta_pl, c_pl, x_mult_z_ext, x_mult_pl_ext, x_nb_ext, x_pl_ext, y, offset_nb, offset_zc, offset_pl_mult, w)
 }
 
-log_lik_profile_fun <- function(gamma_z, gamma_pl, beta_nb, alpha_nb, beta_pl, c_candidates, x_mult_z_ext, x_mult_pl_ext, x_nb_ext, x_pl_ext, y, offset_nb, offset_zc, offset_pl_mult) {
-    .Call(`_evinf_log_lik_profile_fun`, gamma_z, gamma_pl, beta_nb, alpha_nb, beta_pl, c_candidates, x_mult_z_ext, x_mult_pl_ext, x_nb_ext, x_pl_ext, y, offset_nb, offset_zc, offset_pl_mult)
+log_lik_profile_fun <- function(gamma_z, gamma_pl, beta_nb, alpha_nb, beta_pl, c_candidates, x_mult_z_ext, x_mult_pl_ext, x_nb_ext, x_pl_ext, y, offset_nb, offset_zc, offset_pl_mult, w) {
+    .Call(`_evinf_log_lik_profile_fun`, gamma_z, gamma_pl, beta_nb, alpha_nb, beta_pl, c_candidates, x_mult_z_ext, x_mult_pl_ext, x_nb_ext, x_pl_ext, y, offset_nb, offset_zc, offset_pl_mult, w)
 }
 
-update_bfgs_fun <- function(gamma_z_in, gamma_pl_in, beta_nb_in, alpha_nb_in, beta_pl_in, c_pl, x_mult_z_ext, x_mult_pl_ext, x_nb_ext, x_pl_ext, y, max_upd_par, no_m_bfgs_steps, offset_nb, offset_zc, offset_pl_mult, exact_pl = FALSE) {
-    .Call(`_evinf_update_bfgs_fun`, gamma_z_in, gamma_pl_in, beta_nb_in, alpha_nb_in, beta_pl_in, c_pl, x_mult_z_ext, x_mult_pl_ext, x_nb_ext, x_pl_ext, y, max_upd_par, no_m_bfgs_steps, offset_nb, offset_zc, offset_pl_mult, exact_pl)
+update_bfgs_fun <- function(gamma_z_in, gamma_pl_in, beta_nb_in, alpha_nb_in, beta_pl_in, c_pl, x_mult_z_ext, x_mult_pl_ext, x_nb_ext, x_pl_ext, y, max_upd_par, no_m_bfgs_steps, offset_nb, offset_zc, offset_pl_mult, w, exact_pl = FALSE) {
+    .Call(`_evinf_update_bfgs_fun`, gamma_z_in, gamma_pl_in, beta_nb_in, alpha_nb_in, beta_pl_in, c_pl, x_mult_z_ext, x_mult_pl_ext, x_nb_ext, x_pl_ext, y, max_upd_par, no_m_bfgs_steps, offset_nb, offset_zc, offset_pl_mult, w, exact_pl)
 }