+++ src/RcppExports.cpp
@@ -138,8 +138,8 @@ BEGIN_RCPP
 END_RCPP
 }
 // log_lik_fun
-double log_lik_fun(arma::vec gamma_z, arma::vec gamma_pl, arma::vec beta_nb, double alpha_nb, arma::vec beta_pl, double c_pl, arma::mat x_mult_z_ext, arma::mat x_mult_pl_ext, arma::mat x_nb_ext, arma::mat x_pl_ext, arma::vec y, arma::vec offset_nb, arma::vec offset_zc, arma::vec offset_pl_mult);
-RcppExport SEXP _evinf_log_lik_fun(SEXP gamma_zSEXP, SEXP gamma_plSEXP, SEXP beta_nbSEXP, SEXP alpha_nbSEXP, SEXP beta_plSEXP, SEXP c_plSEXP, SEXP x_mult_z_extSEXP, SEXP x_mult_pl_extSEXP, SEXP x_nb_extSEXP, SEXP x_pl_extSEXP, SEXP ySEXP, SEXP offset_nbSEXP, SEXP offset_zcSEXP, SEXP offset_pl_multSEXP) {
+double log_lik_fun(arma::vec gamma_z, arma::vec gamma_pl, arma::vec beta_nb, double alpha_nb, arma::vec beta_pl, double c_pl, arma::mat x_mult_z_ext, arma::mat x_mult_pl_ext, arma::mat x_nb_ext, arma::mat x_pl_ext, arma::vec y, arma::vec offset_nb, arma::vec offset_zc, arma::vec offset_pl_mult, arma::vec w);
+RcppExport SEXP _evinf_log_lik_fun(SEXP gamma_zSEXP, SEXP gamma_plSEXP, SEXP beta_nbSEXP, SEXP alpha_nbSEXP, SEXP beta_plSEXP, SEXP c_plSEXP, SEXP x_mult_z_extSEXP, SEXP x_mult_pl_extSEXP, SEXP x_nb_extSEXP, SEXP x_pl_extSEXP, SEXP ySEXP, SEXP offset_nbSEXP, SEXP offset_zcSEXP, SEXP offset_pl_multSEXP, SEXP wSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -157,13 +157,14 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< arma::vec >::type offset_nb(offset_nbSEXP);
     Rcpp::traits::input_parameter< arma::vec >::type offset_zc(offset_zcSEXP);
     Rcpp::traits::input_parameter< arma::vec >::type offset_pl_mult(offset_pl_multSEXP);
-    rcpp_result_gen = Rcpp::wrap(log_lik_fun(gamma_z, gamma_pl, beta_nb, alpha_nb, beta_pl, c_pl, x_mult_z_ext, x_mult_pl_ext, x_nb_ext, x_pl_ext, y, offset_nb, offset_zc, offset_pl_mult));
+    Rcpp::traits::input_parameter< arma::vec >::type w(wSEXP);
+    rcpp_result_gen = Rcpp::wrap(log_lik_fun(gamma_z, gamma_pl, beta_nb, alpha_nb, beta_pl, c_pl, x_mult_z_ext, x_mult_pl_ext, x_nb_ext, x_pl_ext, y, offset_nb, offset_zc, offset_pl_mult, w));
     return rcpp_result_gen;
 END_RCPP
 }
 // log_lik_profile_fun
-arma::vec log_lik_profile_fun(arma::vec gamma_z, arma::vec gamma_pl, arma::vec beta_nb, double alpha_nb, arma::vec beta_pl, arma::vec c_candidates, arma::mat x_mult_z_ext, arma::mat x_mult_pl_ext, arma::mat x_nb_ext, arma::mat x_pl_ext, arma::vec y, arma::vec offset_nb, arma::vec offset_zc, arma::vec offset_pl_mult);
-RcppExport SEXP _evinf_log_lik_profile_fun(SEXP gamma_zSEXP, SEXP gamma_plSEXP, SEXP beta_nbSEXP, SEXP alpha_nbSEXP, SEXP beta_plSEXP, SEXP c_candidatesSEXP, SEXP x_mult_z_extSEXP, SEXP x_mult_pl_extSEXP, SEXP x_nb_extSEXP, SEXP x_pl_extSEXP, SEXP ySEXP, SEXP offset_nbSEXP, SEXP offset_zcSEXP, SEXP offset_pl_multSEXP) {
+arma::vec log_lik_profile_fun(arma::vec gamma_z, arma::vec gamma_pl, arma::vec beta_nb, double alpha_nb, arma::vec beta_pl, arma::vec c_candidates, arma::mat x_mult_z_ext, arma::mat x_mult_pl_ext, arma::mat x_nb_ext, arma::mat x_pl_ext, arma::vec y, arma::vec offset_nb, arma::vec offset_zc, arma::vec offset_pl_mult, arma::vec w);
+RcppExport SEXP _evinf_log_lik_profile_fun(SEXP gamma_zSEXP, SEXP gamma_plSEXP, SEXP beta_nbSEXP, SEXP alpha_nbSEXP, SEXP beta_plSEXP, SEXP c_candidatesSEXP, SEXP x_mult_z_extSEXP, SEXP x_mult_pl_extSEXP, SEXP x_nb_extSEXP, SEXP x_pl_extSEXP, SEXP ySEXP, SEXP offset_nbSEXP, SEXP offset_zcSEXP, SEXP offset_pl_multSEXP, SEXP wSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -181,13 +182,14 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< arma::vec >::type offset_nb(offset_nbSEXP);
     Rcpp::traits::input_parameter< arma::vec >::type offset_zc(offset_zcSEXP);
     Rcpp::traits::input_parameter< arma::vec >::type offset_pl_mult(offset_pl_multSEXP);
-    rcpp_result_gen = Rcpp::wrap(log_lik_profile_fun(gamma_z, gamma_pl, beta_nb, alpha_nb, beta_pl, c_candidates, x_mult_z_ext, x_mult_pl_ext, x_nb_ext, x_pl_ext, y, offset_nb, offset_zc, offset_pl_mult));
+    Rcpp::traits::input_parameter< arma::vec >::type w(wSEXP);
+    rcpp_result_gen = Rcpp::wrap(log_lik_profile_fun(gamma_z, gamma_pl, beta_nb, alpha_nb, beta_pl, c_candidates, x_mult_z_ext, x_mult_pl_ext, x_nb_ext, x_pl_ext, y, offset_nb, offset_zc, offset_pl_mult, w));
     return rcpp_result_gen;
 END_RCPP
 }
 // update_bfgs_fun
-List update_bfgs_fun(arma::vec gamma_z_in, arma::vec gamma_pl_in, arma::vec beta_nb_in, double alpha_nb_in, arma::vec beta_pl_in, double c_pl, arma::mat x_mult_z_ext, arma::mat x_mult_pl_ext, arma::mat x_nb_ext, arma::mat x_pl_ext, arma::vec y, double max_upd_par, int no_m_bfgs_steps, arma::vec offset_nb, arma::vec offset_zc, arma::vec offset_pl_mult, bool exact_pl);
-RcppExport SEXP _evinf_update_bfgs_fun(SEXP gamma_z_inSEXP, SEXP gamma_pl_inSEXP, SEXP beta_nb_inSEXP, SEXP alpha_nb_inSEXP, SEXP beta_pl_inSEXP, SEXP c_plSEXP, SEXP x_mult_z_extSEXP, SEXP x_mult_pl_extSEXP, SEXP x_nb_extSEXP, SEXP x_pl_extSEXP, SEXP ySEXP, SEXP max_upd_parSEXP, SEXP no_m_bfgs_stepsSEXP, SEXP offset_nbSEXP, SEXP offset_zcSEXP, SEXP offset_pl_multSEXP, SEXP exact_plSEXP) {
+List update_bfgs_fun(arma::vec gamma_z_in, arma::vec gamma_pl_in, arma::vec beta_nb_in, double alpha_nb_in, arma::vec beta_pl_in, double c_pl, arma::mat x_mult_z_ext, arma::mat x_mult_pl_ext, arma::mat x_nb_ext, arma::mat x_pl_ext, arma::vec y, double max_upd_par, int no_m_bfgs_steps, arma::vec offset_nb, arma::vec offset_zc, arma::vec offset_pl_mult, arma::vec w, bool exact_pl);
+RcppExport SEXP _evinf_update_bfgs_fun(SEXP gamma_z_inSEXP, SEXP gamma_pl_inSEXP, SEXP beta_nb_inSEXP, SEXP alpha_nb_inSEXP, SEXP beta_pl_inSEXP, SEXP c_plSEXP, SEXP x_mult_z_extSEXP, SEXP x_mult_pl_extSEXP, SEXP x_nb_extSEXP, SEXP x_pl_extSEXP, SEXP ySEXP, SEXP max_upd_parSEXP, SEXP no_m_bfgs_stepsSEXP, SEXP offset_nbSEXP, SEXP offset_zcSEXP, SEXP offset_pl_multSEXP, SEXP wSEXP, SEXP exact_plSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -207,8 +209,9 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< arma::vec >::type offset_nb(offset_nbSEXP);
     Rcpp::traits::input_parameter< arma::vec >::type offset_zc(offset_zcSEXP);
     Rcpp::traits::input_parameter< arma::vec >::type offset_pl_mult(offset_pl_multSEXP);
+    Rcpp::traits::input_parameter< arma::vec >::type w(wSEXP);
     Rcpp::traits::input_parameter< bool >::type exact_pl(exact_plSEXP);
-    rcpp_result_gen = Rcpp::wrap(update_bfgs_fun(gamma_z_in, gamma_pl_in, beta_nb_in, alpha_nb_in, beta_pl_in, c_pl, x_mult_z_ext, x_mult_pl_ext, x_nb_ext, x_pl_ext, y, max_upd_par, no_m_bfgs_steps, offset_nb, offset_zc, offset_pl_mult, exact_pl));
+    rcpp_result_gen = Rcpp::wrap(update_bfgs_fun(gamma_z_in, gamma_pl_in, beta_nb_in, alpha_nb_in, beta_pl_in, c_pl, x_mult_z_ext, x_mult_pl_ext, x_nb_ext, x_pl_ext, y, max_upd_par, no_m_bfgs_steps, offset_nb, offset_zc, offset_pl_mult, w, exact_pl));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -223,9 +226,9 @@ static const R_CallMethodDef CallEntries[] = {
     {"_evinf_d2elldbeta2_pl_i_fun_approx", (DL_FUNC) &_evinf_d2elldbeta2_pl_i_fun_approx, 4},
     {"_evinf_delldbeta_pl_i_fun_exact", (DL_FUNC) &_evinf_delldbeta_pl_i_fun_exact, 4},
     {"_evinf_d2elldbeta2_pl_i_fun_exact", (DL_FUNC) &_evinf_d2elldbeta2_pl_i_fun_exact, 4},
-    {"_evinf_log_lik_fun", (DL_FUNC) &_evinf_log_lik_fun, 14},
-    {"_evinf_log_lik_profile_fun", (DL_FUNC) &_evinf_log_lik_profile_fun, 14},
-    {"_evinf_update_bfgs_fun", (DL_FUNC) &_evinf_update_bfgs_fun, 17},
+    {"_evinf_log_lik_fun", (DL_FUNC) &_evinf_log_lik_fun, 15},
+    {"_evinf_log_lik_profile_fun", (DL_FUNC) &_evinf_log_lik_profile_fun, 15},
+    {"_evinf_update_bfgs_fun", (DL_FUNC) &_evinf_update_bfgs_fun, 18},
     {NULL, NULL, 0}
 };